+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-## 0.1.0 (August 29th, 2018)
+## 0.2.0 (October 10th, 2018)
+### Added
+- Docker image tagging now accepts a suffix to append to symver tags
+
+### Changed
+- Multiple release assets can now be specified 
+
+## 0.1.0 (October 10th, 2018)
 ### Changed
 - Release asset is now optional
 - Update README with simple usage instructions